Move model conversions to the domain package (renamed from customer) Rename serialisers to remove version
