+++ src/main/scala/com/ovoenergy/orchestration/kafka/IssueOrchestratedComm.scala
@@ -1,7 +1,7 @@
 package com.ovoenergy.orchestration.kafka
 
 import com.ovoenergy.comms.model.{CustomerProfile, TriggeredV3}
-import com.ovoenergy.orchestration.domain.customer.ContactInfo
+import com.ovoenergy.orchestration.domain.ContactInfo
 import org.apache.kafka.clients.producer.RecordMetadata
 
 import scala.concurrent.Future
