+++ src/main/scala/com/ovoenergy/orchestration/profile/CustomerProfiler.scala
@@ -1,11 +1,10 @@
 package com.ovoenergy.orchestration.profile
 
-import com.ovoenergy.orchestration.domain.customer.CustomerProfile
 import okhttp3.{HttpUrl, Request, Response}
 import com.ovoenergy.orchestration.http.JsonDecoding._
 import com.ovoenergy.orchestration.logging.LoggingWithMDC
 import com.ovoenergy.orchestration.retry.Retry.{Failed, RetryConfig}
-import com.ovoenergy.orchestration.domain.customer._
+import com.ovoenergy.orchestration.domain._
 import io.circe.generic.auto._
 import cats.syntax.either._
 import com.ovoenergy.comms.model.ProfileRetrievalFailed
